Organize `eth_client` modules
