+++ core/lib/eth_client/src/clients/mod.rs
@@ -1,2 +1,10 @@
-pub mod http;
-pub mod mock;
+//! Various Ethereum client implementations.
+
+mod generic;
+mod http;
+mod mock;
+
+pub use self::{
+    http::{PKSigningClient, QueryClient, SigningClient},
+    mock::MockEthereum,
+};
