+++ core/lib/zksync_core/src/eth_sender/error.rs
@@ -1,10 +1,9 @@
-use zksync_eth_client::types;
 use zksync_types::web3::contract;
 
 #[derive(Debug, thiserror::Error)]
 pub enum ETHSenderError {
     #[error("Ethereum gateway Error {0}")]
-    EthereumGateWayError(#[from] types::Error),
+    EthereumGateWayError(#[from] zksync_eth_client::Error),
     #[error("Token parsing Error: {0}")]
     ParseError(#[from] contract::Error),
 }
