+++ core/lib/zksync_core/src/eth_sender/tests.rs
@@ -8,7 +8,7 @@ use zksync_config::{
 };
 use zksync_contracts::BaseSystemContractsHashes;
 use zksync_dal::{ConnectionPool, StorageProcessor};
-use zksync_eth_client::{clients::mock::MockEthereum, EthInterface};
+use zksync_eth_client::{clients::MockEthereum, EthInterface};
 use zksync_object_store::ObjectStoreFactory;
 use zksync_types::{
     aggregated_operations::{
